user_service/requirements.txt
fastapi
uvicorn
redis
python-dotenv
sqlalchemy
pydantic[email]
psycopg2-binary  